fix: change disabled approach
